test: add platform dependent tests
